--- gpu-requirements.in
@@ -1,8 +1,0 @@
--r requirements.txt
-autoawq
-sentence-transformers
-vllm
-xformers
-
---extra-index-url https://download.pytorch.org/whl/cu121
-torch
