+++ requirements/base.in
@@ -1,0 +1,14 @@
+fastapi
+gensim>=4.3
+huggingface_hub
+httpx
+nltk>=3.8
+spacy>=3.0,<4.0
+https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
+supabase
+protobuf==3.20.*
+pycld2>=0.40
+pydantic>=1.2.4,<2.0.0
+transformers>=4.31.0
+uvicorn
+youtube_transcript_api>=0.6.0
