+++ requirements/base.txt
@@ -1,0 +1,239 @@
+#
+# This file is autogenerated by pip-compile with Python 3.10
+# by the following command:
+#
+#    pip-compile requirements/base.in
+#
+anyio==3.7.1
+    # via
+    #   fastapi
+    #   httpcore
+    #   starlette
+blis==0.7.11
+    # via thinc
+catalogue==2.0.10
+    # via
+    #   spacy
+    #   srsly
+    #   thinc
+certifi==2023.11.17
+    # via
+    #   httpcore
+    #   httpx
+    #   requests
+charset-normalizer==3.3.2
+    # via requests
+click==8.1.7
+    # via
+    #   nltk
+    #   typer
+    #   uvicorn
+cloudpathlib==0.16.0
+    # via weasel
+confection==0.1.4
+    # via
+    #   thinc
+    #   weasel
+cymem==2.0.8
+    # via
+    #   preshed
+    #   spacy
+    #   thinc
+deprecation==2.1.0
+    # via postgrest
+en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
+    # via -r requirements/base.in
+exceptiongroup==1.2.0
+    # via anyio
+fastapi==0.105.0
+    # via -r requirements/base.in
+filelock==3.13.1
+    # via
+    #   huggingface-hub
+    #   transformers
+fsspec==2023.12.2
+    # via huggingface-hub
+gensim==4.3.2
+    # via -r requirements/base.in
+gotrue==2.1.0
+    # via supabase
+h11==0.14.0
+    # via
+    #   httpcore
+    #   uvicorn
+httpcore==0.17.3
+    # via httpx
+httpx==0.24.1
+    # via
+    #   -r requirements/base.in
+    #   gotrue
+    #   postgrest
+    #   storage3
+    #   supabase
+    #   supafunc
+huggingface-hub==0.19.4
+    # via
+    #   -r requirements/base.in
+    #   tokenizers
+    #   transformers
+idna==3.6
+    # via
+    #   anyio
+    #   httpx
+    #   requests
+jinja2==3.1.2
+    # via spacy
+joblib==1.3.2
+    # via nltk
+langcodes==3.3.0
+    # via spacy
+markupsafe==2.1.3
+    # via jinja2
+murmurhash==1.0.10
+    # via
+    #   preshed
+    #   spacy
+    #   thinc
+nltk==3.8.1
+    # via -r requirements/base.in
+numpy==1.26.2
+    # via
+    #   blis
+    #   gensim
+    #   scipy
+    #   spacy
+    #   thinc
+    #   transformers
+packaging==23.2
+    # via
+    #   deprecation
+    #   huggingface-hub
+    #   spacy
+    #   thinc
+    #   transformers
+    #   weasel
+postgrest==0.13.0
+    # via supabase
+preshed==3.0.9
+    # via
+    #   spacy
+    #   thinc
+protobuf==3.20.3
+    # via -r requirements/base.in
+pycld2==0.41
+    # via -r requirements/base.in
+pydantic==1.10.13
+    # via
+    #   -r requirements/base.in
+    #   confection
+    #   fastapi
+    #   gotrue
+    #   postgrest
+    #   spacy
+    #   thinc
+    #   weasel
+python-dateutil==2.8.2
+    # via
+    #   realtime
+    #   storage3
+pyyaml==6.0.1
+    # via
+    #   huggingface-hub
+    #   transformers
+realtime==1.0.2
+    # via supabase
+regex==2023.10.3
+    # via
+    #   nltk
+    #   transformers
+requests==2.31.0
+    # via
+    #   huggingface-hub
+    #   spacy
+    #   transformers
+    #   weasel
+    #   youtube-transcript-api
+safetensors==0.4.1
+    # via transformers
+scipy==1.11.4
+    # via gensim
+six==1.16.0
+    # via python-dateutil
+smart-open==6.4.0
+    # via
+    #   gensim
+    #   spacy
+    #   weasel
+sniffio==1.3.0
+    # via
+    #   anyio
+    #   httpcore
+    #   httpx
+spacy==3.7.2
+    # via
+    #   -r requirements/base.in
+    #   en-core-web-sm
+spacy-legacy==3.0.12
+    # via spacy
+spacy-loggers==1.0.5
+    # via spacy
+srsly==2.4.8
+    # via
+    #   confection
+    #   spacy
+    #   thinc
+    #   weasel
+starlette==0.27.0
+    # via fastapi
+storage3==0.7.0
+    # via supabase
+strenum==0.4.15
+    # via postgrest
+supabase==2.3.0
+    # via -r requirements/base.in
+supafunc==0.3.1
+    # via supabase
+thinc==8.2.2
+    # via spacy
+tokenizers==0.15.0
+    # via transformers
+tqdm==4.66.1
+    # via
+    #   huggingface-hub
+    #   nltk
+    #   spacy
+    #   transformers
+transformers==4.36.1
+    # via -r requirements/base.in
+typer==0.9.0
+    # via
+    #   spacy
+    #   weasel
+typing-extensions==4.9.0
+    # via
+    #   cloudpathlib
+    #   fastapi
+    #   huggingface-hub
+    #   pydantic
+    #   realtime
+    #   storage3
+    #   typer
+    #   uvicorn
+urllib3==2.1.0
+    # via requests
+uvicorn==0.24.0.post1
+    # via -r requirements/base.in
+wasabi==1.1.2
+    # via
+    #   spacy
+    #   thinc
+    #   weasel
+weasel==0.3.4
+    # via spacy
+websockets==11.0.3
+    # via realtime
+youtube-transcript-api==0.6.1
+    # via -r requirements/base.in
+
+# The following packages are considered to be unsafe in a requirements file:
+# setuptools
