+++ src/main.py
@@ -3,17 +3,17 @@ import os
 from fastapi import FastAPI, HTTPException, Response
 from fastapi.middleware.cors import CORSMiddleware
 
-from models.summary import SummaryInput, SummaryResults
-from models.answer import AnswerInput, AnswerResults
-from models.embedding import ChunkInput, RetrievalInput, RetrievalResults
-from models.chat import ChatInput, ChatResult
-from models.transcript import TranscriptInput, TranscriptResults
+from .models.summary import SummaryInput, SummaryResults
+from .models.answer import AnswerInput, AnswerResults
+from .models.embedding import ChunkInput, RetrievalInput, RetrievalResults
+from .models.chat import ChatInput, ChatResult
+from .models.transcript import TranscriptInput, TranscriptResults
 
-from src.summary_eval_supabase import summary_score_supabase
-from src.summary_eval import summary_score
-from src.answer_eval_supabase import answer_score_supabase
-from src.answer_eval import answer_score
-from src.transcript import transcript_generate
+from .summary_eval_supabase import summary_score_supabase
+from .summary_eval import summary_score
+from .answer_eval_supabase import answer_score_supabase
+from .answer_eval import answer_score
+from .transcript import transcript_generate
 
 description = """
 Welcome to iTELL AI, a REST API for intelligent textbooks. iTELL AI provides the following principal features:
@@ -80,15 +80,12 @@ def gpu() -> dict[str, str | bool]:
         return {"message": torch.cuda.is_available()}
 
 
-@app.post(
-    "/score/summary",
-    description=(
-        "Score a summary."
-        " Requires a textbook name if the textbook content is on SupaBase."
-        " Requires a page_slug if the textbook content is in our Strapi CMS."
-    ),
-)
+@app.post("/score/summary")
 async def score_summary(input_body: SummaryInput) -> SummaryResults:
+    """Score a summary.
+    Requires a textbook name if the textbook content is on SupaBase.
+    Requires a page_slug if the textbook content is in our Strapi CMS.
+    """
     input_body = SummaryInput.parse_obj(input_body)
     if input_body.textbook_name:  # supabase requires textbook_name (deprecated)
         return await summary_score_supabase(input_body)
@@ -96,15 +93,12 @@ async def score_summary(input_body: SummaryInput) -> SummaryResults:
         return await summary_score(input_body)
 
 
-@app.post(
-    "/score/answer",
-    description=(
-        "Score a constructed response item."
-        " Requires a textbook name and location IDs if the textbook content is on SupaBase."
-        " Requires a page_slug and chunk_slug if the textbook content is in our Strapi CMS."
-    ),
-)
+@app.post("/score/answer")
 async def score_answer(input_body: AnswerInput) -> AnswerResults:
+    """Score a constructed response item.
+    Requires a textbook name and location IDs if the textbook content is on SupaBase.
+    Requires a page_slug and chunk_slug if the textbook content is in our Strapi CMS.
+    """
     if input_body.textbook_name:  # supabase requires textbook_name (deprecated)
         return await answer_score_supabase(input_body)
     else:  # Strapi method
